+++ apps/govuk_utils/forms.py
@@ -1,20 +1,75 @@
-from django.forms.extras.widgets import SelectDateWidget
-from django.forms.widgets import NumberInput
+from django import forms
+from django.forms.formsets import BaseFormSet
 
 
-class GovUkDateWidget(SelectDateWidget):
-    """
-    Same as SelectDateWidget, but uses NumberInput rather than Select.
-    """
-    def create_select(self, name, field, value, val, choices):
-        none_value = ""
-        if 'id' in self.attrs:
-            id_ = self.attrs['id']
-        else:
-            id_ = 'id_%s' % name
-        if not self.is_required:
-            choices.insert(0, none_value)
-        local_attrs = self.build_attrs(id=field % id_)
-        s = NumberInput()
-        select_html = s.render(field % name, val, local_attrs)
-        return select_html
+
+class RequiredFormSet(BaseFormSet):
+    def __init__(self, *args, **kwargs):
+        super(RequiredFormSet, self).__init__(*args, **kwargs)
+        for form in self.forms:
+            form.empty_permitted = False
+
+
+class BaseStageForm(forms.Form):
+    def __init__(self, *args, **kwargs):
+        super(BaseStageForm, self).__init__(*args, **kwargs)
+        try:
+            self.data = args[0]
+        except IndexError:
+            self.data = kwargs.get("data", {})
+
+        self.split_form = self.data.get("split_form", None)
+
+        if hasattr(self, "dependencies"):
+            prefix = kwargs.get("prefix", None)
+            self.check_dependencies(self.dependencies, prefix)
+
+    def check_dependencies(self, dependencies_list, prefix=None):
+        """
+        Set the required attribute depending on the dependencies map
+        and the already submitted data
+        """
+        for field, dependency in dependencies_list.items():
+            dependency_field = dependency.get("field", None)
+            dependency_value = dependency.get("value", None)
+            sub_dependencies = dependency.get("dependencies", None)
+
+            """
+            When a form has a prefix, the key in the field is the original,
+            but the key in data is updated. Would there be a nicer way of
+            handling this?
+            """
+            if prefix:
+                dependency_field_data_key = prefix + "-" + dependency_field
+            else:
+                dependency_field_data_key = dependency_field
+
+            if self.fields.get(field, None):
+
+                self.fields[field].required = False
+
+                if self.fields[dependency_field].required:
+                    if self.split_form is None or self.split_form != dependency_field:
+                        if dependency_value and self.data.get(dependency_field_data_key, None) == dependency_value:
+                            self.fields[field].required = True
+                        
+                        if not dependency_value and self.data.get(dependency_field_data_key, None) is not None:
+                            self.fields[field].required = True
+
+                    if sub_dependencies:
+                        self.check_dependencies(sub_dependencies, prefix)
+
+
+class SplitStageForm(BaseStageForm):
+    split_form = forms.CharField(widget=forms.HiddenInput(), required=False)
+
+    split_form_options = {}
+
+    def __init__(self, *args, **kwargs):
+        super(SplitStageForm, self).__init__(*args, **kwargs)
+
+        if self.split_form is None:
+            self.fields["split_form"].initial = self.split_form_options.get("trigger", False)
+
+        if self.split_form_options.get("nojs_only", False):
+            self.fields["split_form"].widget.attrs.update({"class": "nojs-only"})
