+++ apps/plea/forms.py
@@ -1,7 +1,6 @@
 from __future__ import unicode_literals
 
 from django import forms
-from django.forms.formsets import BaseFormSet
 from django.forms.widgets import Textarea, RadioSelect
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
@@ -11,6 +10,8 @@ from .validators import (is_date_in_past,
                          is_date_within_range,
                          is_urn_not_used,
                          is_urn_valid)
+
+from apps.govuk_utils.forms import BaseStageForm, SplitStageForm
 
 from .fields import (ERROR_MESSAGES,
                      DSRadioFieldRenderer,
@@ -24,79 +25,7 @@ YESNO_CHOICES = (
 to_bool = lambda x: x == "True"
 
 
-class RequiredFormSet(BaseFormSet):
-    def __init__(self, *args, **kwargs):
-        super(RequiredFormSet, self).__init__(*args, **kwargs)
-        for form in self.forms:
-            form.empty_permitted = False
-
-
-class BasePleaStepForm(forms.Form):
-    def __init__(self, *args, **kwargs):
-        super(BasePleaStepForm, self).__init__(*args, **kwargs)
-        try:
-            self.data = args[0]
-        except IndexError:
-            self.data = kwargs.get("data", {})
-
-        self.split_form = self.data.get("split_form", None)
-
-        if hasattr(self, "dependencies"):
-            prefix = kwargs.get("prefix", None)
-            self.check_dependencies(self.dependencies, prefix)
-
-    def check_dependencies(self, dependencies_list, prefix=None):
-        """
-        Set the required attribute depending on the dependencies map
-        and the already submitted data
-        """
-        for field, dependency in dependencies_list.items():
-            dependency_field = dependency.get("field", None)
-            dependency_value = dependency.get("value", None)
-            sub_dependencies = dependency.get("dependencies", None)
-
-            """
-            When a form has a prefix, the key in the field is the original,
-            but the key in data is updated. Would there be a nicer way of
-            handling this?
-            """
-            if prefix:
-                dependency_field_data_key = prefix + "-" + dependency_field
-            else:
-                dependency_field_data_key = dependency_field
-
-            if self.fields.get(field, None):
-
-                self.fields[field].required = False
-
-                if self.fields[dependency_field].required:
-                    if self.split_form is None or self.split_form != dependency_field:
-                        if dependency_value and self.data.get(dependency_field_data_key, None) == dependency_value:
-                            self.fields[field].required = True
-                        
-                        if not dependency_value and self.data.get(dependency_field_data_key, None) is not None:
-                            self.fields[field].required = True
-
-                    if sub_dependencies:
-                        self.check_dependencies(sub_dependencies, prefix)
-
-
-class SplitPleaStepForm(BasePleaStepForm):
-    split_form = forms.CharField(widget=forms.HiddenInput(), required=False)
-
-    split_form_options = {}
-
-    def __init__(self, *args, **kwargs):
-        super(SplitPleaStepForm, self).__init__(*args, **kwargs)
-
-        if self.split_form is None:
-            self.fields["split_form"].initial = self.split_form_options.get("trigger", False)
-
-        if self.split_form_options.get("nojs_only", False):
-            self.fields["split_form"].widget.attrs.update({"class": "nojs-only"})
-
-
-class CaseForm(BasePleaStepForm):
+class CaseForm(BaseStageForm):
     PLEA_MADE_BY_CHOICES = (
         ("Defendant", _("The person named in the notice")),
         ("Company representative", _("Pleading on behalf of a company")))
@@ -136,7 +65,7 @@ class CaseForm(BasePleaStepForm):
                                           error_messages={"required": ERROR_MESSAGES["PLEA_MADE_BY_REQUIRED"]})
 
 
-class YourDetailsForm(BasePleaStepForm):
+class YourDetailsForm(BaseStageForm):
     dependencies = {
         "updated_address": {
             "field": "correct_address",
@@ -196,7 +125,7 @@ class YourDetailsForm(BasePleaStepForm):
                                              help_text=_("Starts with letters from your last name."))
 
 
-class CompanyDetailsForm(BasePleaStepForm):
+class CompanyDetailsForm(BaseStageForm):
     dependencies = {
         "updated_address": {
             "field": "correct_address",
@@ -255,7 +184,7 @@ class CompanyDetailsForm(BasePleaStepForm):
                                                      "invalid": ERROR_MESSAGES["CONTACT_NUMBER_INVALID"]})
 
 
-class YourMoneyForm(SplitPleaStepForm):
+class YourMoneyForm(SplitStageForm):
 
     YOU_ARE_CHOICES = (("Employed", _("Employed")),
                        ("Self-employed", _("Self-employed")),
@@ -424,7 +353,7 @@ class YourMoneyForm(SplitPleaStepForm):
                                             error_messages={"required": ERROR_MESSAGES["HARDSHIP_REQUIRED"]})
 
 
-class YourExpensesForm(BasePleaStepForm):
+class YourExpensesForm(BaseStageForm):
     hardship_details = forms.CharField(
         label=_("How would paying a fine cause you serious financial problems?"),
         help_text=_("What should the court consider when deciding on any possible fine?"),
@@ -567,7 +496,7 @@ class YourExpensesForm(BasePleaStepForm):
                         'min_value': ERROR_MESSAGES['OTHER_CHILD_MAINTENANCE_MIN']})
 
 
-class CompanyFinancesForm(SplitPleaStepForm):
+class CompanyFinancesForm(SplitStageForm):
     dependencies = {
         "number_of_employees": {
             "field": "trading_period"
@@ -630,7 +559,7 @@ class CompanyFinancesForm(SplitPleaStepForm):
             self.fields["net_turnover"].error_messages.update({"required": ERROR_MESSAGES["COMPANY_NET_TURNOVER"]})
 
 
-class ConfirmationForm(BasePleaStepForm):
+class ConfirmationForm(BaseStageForm):
     dependencies = {
         "email": {
             "field": "receive_email_updates",
@@ -657,7 +586,7 @@ class ConfirmationForm(BasePleaStepForm):
                                     error_messages={"required": ERROR_MESSAGES["UNDERSTAND_REQUIRED"]})
 
 
-class PleaForm(SplitPleaStepForm):
+class PleaForm(SplitStageForm):
     PLEA_CHOICES = (
         ('guilty', _('Guilty')),
         ('not_guilty', _('Not guilty')),
