+++ apps/plea/stages.py
@@ -8,10 +8,17 @@ from django.utils.translation import ugettext as _
 from django.shortcuts import redirect
 
 from apps.govuk_utils.stages import FormStage
-from email import send_plea_email
-from forms import (CaseForm, YourDetailsForm, CompanyDetailsForm,
-                   PleaForm, YourMoneyForm, YourExpensesForm,
-                   CompanyFinancesForm, ConfirmationForm, RequiredFormSet)
+from apps.govuk_utils.forms import RequiredFormSet
+
+from .email import send_plea_email
+from .forms import (CaseForm,
+                    YourDetailsForm,
+                    CompanyDetailsForm,
+                    PleaForm,
+                    YourMoneyForm,
+                    YourExpensesForm,
+                    CompanyFinancesForm,
+                    ConfirmationForm)
 
 from .fields import ERROR_MESSAGES
 from .models import Court, Case, Offence
